+++ tests_from_past/integration/scollection_count_test.py
@@ -20,4 +20,40 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
 
-VERSION = '2.0.0a7'
+from selene.support.past import config
+from selene.support.past.common.none_object import NoneObject
+from selene.support.past.driver import SeleneDriver
+from tests_from_past.past.acceptance import get_test_driver
+from tests_from_past.integration.helpers import GivenPage
+
+__author__ = 'yashaka'
+
+driver = NoneObject('driver')  # type: SeleneDriver
+GIVEN_PAGE = NoneObject('GivenPage')  # type: GivenPage
+WHEN = GIVEN_PAGE  # type: GivenPage
+original_timeout = config.timeout
+
+
+def setup_module(m):
+    global driver
+    driver = SeleneDriver.wrap(get_test_driver())
+    global GIVEN_PAGE
+    GIVEN_PAGE = GivenPage(driver)
+    global WHEN
+    WHEN = GIVEN_PAGE
+
+
+def teardown_module(m):
+    driver.quit()
+
+
+def test_counts_invisible_tasks():
+    GIVEN_PAGE.opened_empty()
+    elements = driver.all('.will-appear')
+
+    WHEN.load_body('''
+                   <ul>Hello to:
+                       <li class='will-appear'>Bob</li>
+                       <li class='will-appear' style='display:none'>Kate</li>
+                   </ul>''')
+    assert len(elements) == 2