+++ tests_from_past/past/acceptance/bys_test.py
@@ -20,4 +20,27 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
 
-VERSION = '2.0.0a7'
+from tests_from_past.past.acceptance import get_test_driver
+
+__author__ = 'yashaka'
+
+from selene.support.past.browser import *
+from tests_from_past.past.acceptance import given_active
+from selene.support.past.bys import *
+
+
+def setup_module(m):
+    set_driver(get_test_driver())
+
+
+def teardown_module(m):
+    driver().quit()
+
+
+def test_by_text_with_single_and_double_quotes():
+    given_active("""Fred's last name is "Li".""")
+    assert find_element(by_text("""Fred's last name is "Li".""")).is_displayed()
+
+
+def find_element(locator):
+    return driver().find_element(*locator)