+++ tests_from_past/past/acceptance/selene_element_inner_search_should_test.py
@@ -20,4 +20,29 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
 
-VERSION = '2.0.0a7'
+from selene.support.conditions import have
+from selene.support.jquery_style_selectors import s
+from tests_from_past.past.acceptance import get_test_driver
+
+__author__ = 'yashaka'
+
+from selene.support.past.browser import *
+from tests_from_past.past.acceptance import given_active
+
+
+def setup_module(m):
+    set_driver(get_test_driver())
+
+
+def teardown_module(m):
+    driver().quit()
+
+
+def test_search_inner_selement():
+    given_active("a", "b")
+    s("#todo-list").s("li").should(have.exact_text("a"))
+
+
+def test_search_inner_selene_collection():
+    given_active("a", "b")
+    s("#todo-list").all("li").should(have.exact_texts("a", "b"))